+++ Civi/ActionProvider/Utils/SendEmail.php
@@ -142,7 +142,9 @@ class SendEmail {
         $toName = $contact['display_name'];
       }
 
-      \CRM_Utils_Hook::tokenValues($contact, $contact['contact_id'], NULL, $tokens);
+      $tokenValues[$contact['contact_id']] = $contact;
+      \CRM_Utils_Hook::tokenValues($tokenValues, array($contact['contact_id']), NULL, $tokens);
+      $contact = $tokenValues[$contact['contact_id']];
       // call token hook
       $hookTokens = array();
       \CRM_Utils_Hook::tokens($hookTokens);
@@ -163,10 +165,10 @@ class SendEmail {
           }
 
           $$bodyType = \CRM_Utils_Token::replaceDomainTokens($$bodyType, $domain, TRUE, $tokens, TRUE);
-          $$bodyType = \CRM_Utils_Token::replaceHookTokens($$bodyType, $contact, $categories, TRUE);
-          \CRM_Utils_Token::replaceGreetingTokens($$bodyType, $contact, $contact['contact_id']);
           $$bodyType = \CRM_Utils_Token::replaceContactTokens($$bodyType, $contact, FALSE, $tokens, FALSE, TRUE);
           $$bodyType = \CRM_Utils_Token::replaceComponentTokens($$bodyType, $contact, $tokens, TRUE);
+          $$bodyType = \CRM_Utils_Token::replaceHookTokens($$bodyType, $contact, $categories, TRUE);
+          \CRM_Utils_Token::replaceGreetingTokens($$bodyType, $contact, $contact['contact_id']);
         }
       }
       $html = $body_html;
@@ -179,10 +181,11 @@ class SendEmail {
       }
 
       // do replacements in message subject
-      $messageSubject = \CRM_Utils_Token::replaceContactTokens($subject, $contact, false, $tokens);
-      $messageSubject = \CRM_Utils_Token::replaceDomainTokens($messageSubject, $domain, true, $tokens);
+      $messageSubject = \CRM_Utils_Token::replaceDomainTokens($subject, $domain, true, $tokens);
+      $messageSubject = \CRM_Utils_Token::replaceContactTokens($messageSubject, $contact, false, $tokens);
       $messageSubject = \CRM_Utils_Token::replaceComponentTokens($messageSubject, $contact, $tokens, true);
       $messageSubject = \CRM_Utils_Token::replaceHookTokens($messageSubject, $contact, $categories, true);
+      \CRM_Utils_Token::replaceGreetingTokens($messageSubject, $contact, $contact['contact_id']);
 
       if (defined('CIVICRM_MAIL_SMARTY') && CIVICRM_MAIL_SMARTY) {
         $messageSubject = $smarty->fetch("string:{$messageSubject}");